aarch64: support PAC and BTI
Enable Pointer Authentication Codes (PAC) and Branch Target
Identification (BTI) support for ARM 64 targets.

PAC works by signing the LR with either an A key or B key and verifying
the return address. There are quite a few instructions capable of doing
this, however, the Linux ARM ABI is to use hint compatible instructions
that can be safely NOP'd on older hardware and can be assembled and
linked with older binutils. This limits the instruction set to paciasp,
pacibsp, autiasp and autibsp. Instructions prefixed with pac are for
signing and instructions prefixed with aut are for signing. Both
instructions are then followed with an a or b to indicate which signing
key they are using. The keys can be controlled using
-mbranch-protection=pac-ret for the A key and
-mbranch-protection=pac-ret+b-key for the B key.

BTI works by marking all call and jump positions with bti c and bti
j instructions. If execution control transfers to an instruction other
than a BTI instruction, the execution is killed via SIGILL. Note that
to remove one instruction, the aforementioned pac instructions will
also work as a BTI landing pad for bti c usages.

For BTI to work, all object files linked for a unit of execution,
whether an executable or a library must have the GNU Notes section of
the ELF file marked to indicate BTI support. This is so loader/linkers
can apply the proper permission bits (PROT_BRI) on the memory region.

PAC can also be annotated in the GNU ELF notes section, but it's not
required for enablement, as interleaved PAC and non-pac code works as
expected since it's the callee that performs all the checking. The
linker follows the same rules as BTI for discarding the PAC flag from
the GNU Notes section.

Testing was done under the following CFLAGS and CXXFLAGS for all
combinations:
1. -mbranch-protection=none
2. -mbranch-protection=standard
3. -mbranch-protection=pac-ret
4. -mbranch-protection=pac-ret+b-key
5. -mbranch-protection=bti

Additional Notes:
MPN was handled differently then the standard approach of all PROLOGUES
getting a SIGN_LR macro. This is becuase MPN does not make use of
saving the x30, aka the link regeister (LR), to the stack in almost all
instances. However, some functions do, and they were explicitly handled.
This not only avoids the cost of the operations to sign and verify the
LR but also handles instances where branches are taken to labels where
indirect branches are used over branch and link to optimize the assembly.

Also, within the configure.ac are a myriad of options for different
architectures, chipsets, ABIs, etc. To compound that, additional
architecture specifiec features could be enabled with in CFLAGS that
needs to be respected in-order to get a correct output. For instance in
aarch64, the PAC and BTI instructions need to be output in the generated
assembly as well as the GNU notes section added to the ELF output to get
those security features. Hacking it into the configure options seems
baroque, especially considering that distro packaging will often just
set a set of CFLAGS to be respected and move on and that's most users
would expect. Taking this all into consideration, allowing for a per
architecture script that can be executed to generate additional m4
allows for internal definitions, like in the PAC case, to be exposed, or
any multitude of options if other archs need somethng like this. This
introduces the variable gen_path_m4 that arch's can set to the script of
their choosing to generate whatever m4 they need that is prepended to
the m4 generation command after the defines.

Signed-off-by: Bill Roberts <bill.roberts@arm.com>